+++ CHANGELOG.md
@@ -1,4 +1,6 @@
 # Changelog
+## 2024-06-10
+- Add identifier to check unique transactions.
 ## 2024-06-09
 - Add testing.
 - Fix currency extraction.